# Reduce overhead of kwargs for no-arg methods and blocks