--- a/core/src/main/java/org/jruby/ir/IRBuilder.java
+++ b/core/src/main/java/org/jruby/ir/IRBuilder.java
@@ -3129,6 +3129,13 @@ public class IRBuilder {
      */
     public void receiveArgs(final ArgsNode argsNode) {
         Signature signature = scope.getStaticScope().getSignature();
+
+        if (signature.equals(Signature.NO_ARGUMENTS)) {
+            addInstr(new CheckArityInstr(0, 0, false, -1, UndefinedValue.UNDEFINED));
+            receiveBlockArg(argsNode);
+            return;
+        }
+
         Variable keywords = addResultInstr(new ReceiveKeywordsInstr(temp(), signature.hasRest(), argsNode.hasKwargs()));
 
         KeywordRestArgNode keyRest = argsNode.getKeyRest();